simple CRUD API completed

+++ src/main/java/com/schoolmanagementsystem/schoolmanagementsystem/models/Student.java
@@ -6,18 +6,21 @@ import jakarta.persistence.*;
 @Table(name = "students")
 public class Student {
     @Id
-    private int id;
+    private Long id;
 
     private String name;
     private Integer age;
     private Integer rollNo;
 
-    public Integer getId(){
+
+    public Long getId() {
         return id;
     }
-    public void setId(Integer id){
+
+    public void setId(Long id) {
         this.id = id;
     }
+
     public String getName() {
         return name;
     }
